+++ docs/sphinx/source/resources/pages/changes.rst
@@ -1,6 +1,14 @@
 Changes
 =======
 
+2.0.5 - Stable
+--------------
+
+| **Manager** 2.0.5 - Stable - Milestone: https://github.com/KelSolaar/Manager/issues?milestone=4&state=closed
+| **Foundations** 2.1.0 - Stable - Milestone: https://github.com/KelSolaar/Foundations/issues?milestone=8&page=1&state=closed
+
+-  Reformatted package docstrings.
+
 2.0.4 - Stable
 --------------